[WebGL] Add ability to override default priority for WebGL content https://bugs.webkit.org/show_bug.cgi?id=257282 <radar://104710441>
Reviewed by NOBODY (OOPS!).

Add ability to override the priority of the command queues.

Use swizzling so we don't need to add SPI to Angle.

Also limit the change to running WebKit.WebContent or WebKit.GPU processes,
as many other processes link WebCore and we don't want to set their queues
to low priority.

* Source/WebCore/PAL/pal/spi/cocoa/MetalSPI.h:
I only reference low priority directly so that is the only value I staged.

* Source/WebCore/platform/graphics/cocoa/GraphicsContextGLCocoa.mm:
(+[_MTLCommandQueue initialize]):
(-[_MTLCommandQueue WebKit_initWithDevice:]):
Use swizzling so we don't need to add SPI to Angle.