[GPUP] Tune minimumRenderingUpdateCountToKeepFontAlive https://bugs.webkit.org/show_bug.cgi?id=257288 rdar://109580010
Reviewed by NOBODY (OOPS!).

Currently, web font objects created in the web process live for the lifetime of the
document. When we need to render glyphs with a font, we send the font to the GPU
process. We don't want to pay the memory cost of every font twice (once for the web
process, and once for the GPU process), so we instead treat the GPU process as a
cache. Fonts in the GPU process live until they are unused for
minimumRenderingUpdateCountToKeepFontAlive number of rendering updates. At that
point, they are destroyed (and if they are needed again, are resent from the web
process).

If we set this threshold too low, then we end up spending lots of power recreating
lots of fonts for every rendering update. If we set the threshold too high, we
increase memory use for all the duplicated font objects that are alive.

I analyzed our power benchmark, and discovered that, with the current threshold
value, we end up sending 2.15x as many fonts as we need to send. This is showing up
on our power benchmark as a potential optimization opportunity. If we increase the
threshold to 41, we end up sending every font exactly once, which is optimal from
a power perspective.

* Source/WebKit/WebProcess/GPU/graphics/RemoteResourceCacheProxy.cpp:
(WebKit::RemoteResourceCacheProxy::finalizeRenderingUpdateForFonts):