AX: Various attributedStringSetProperty methods unnecessarily call -[NSMutableAttributedString removeAttribute:range:] https://bugs.webkit.org/show_bug.cgi?id=257287 rdar://problem/109798594
Reviewed by Chris Fleizach.

When we call attributedStringCreate, we always start with a fresh
attributed string. Then as we build the attributed string with the
various attributedStringSetFoo methods, we add the attribute to a given
range if necessary, and remove the attribute otherwise. But this removal
is completely unnecessary since we know we never added it in the first
place because the attributed string is brand new.

This saves 4.6k of 38k total samples on an academic journal webpage with
a lot of text (and therefore a lot of attributed strings that need creating).

* Source/WebCore/accessibility/cocoa/AccessibilityObjectCocoa.mm:
(WebCore::attributedStringSetNumber):
(WebCore::attributedStringSetFont):
* Source/WebCore/accessibility/mac/AccessibilityObjectMac.mm:
(WebCore::attributedStringSetColor):
(WebCore::attributedStringSetStyle):
(WebCore::attributedStringSetHeadingLevel):
(WebCore::attributedStringSetBlockquoteLevel):
(WebCore::attributedStringSetExpandedText):
(WebCore::attributedStringSetElement):

Canonical link: https://commits.webkit.org/264549@main