[WGSL] Expand accepted syntax in for loops https://bugs.webkit.org/show_bug.cgi?id=257320 rdar://109827463
Reviewed by Mike Wyrzykowski.

Accept a "variable-updating-statement" in for-init and for-update, as per the
reference grammar. After this patch, the last missing piece of syntax for
for-loops is accepting calls in for-init and for-update, but that requires
refactoring how we parse function calls first.

* Source/WebGPU/WGSL/Metal/MetalFunctionWriter.cpp:
(WGSL::Metal::FunctionDefinitionWriter::visit):
* Source/WebGPU/WGSL/Parser.cpp:
(WGSL::Parser<Lexer>::parseStatement):
(WGSL::Parser<Lexer>::parseForStatement):
(WGSL::Parser<Lexer>::parseVariableUpdatingStatement):
* Source/WebGPU/WGSL/ParserPrivate.h:
* Source/WebGPU/WGSL/TypeCheck.cpp:
(WGSL::TypeChecker::visit):
* Source/WebGPU/WGSL/tests/invalid/for.wgsl:
* Source/WebGPU/WGSL/tests/valid/for.wgsl:

Canonical link: https://commits.webkit.org/264577@main