[WebGPU] DeviceImpl::createBindGroup may access re-allocated memory https://bugs.webkit.org/show_bug.cgi?id=257192 <radar://109704196>
Reviewed by Myles C. Maxfield.

It is unsafe to store pointers to elements in a vector when the size is changing.

Address this by using SegmentedVector with a segment size of 1 as we expect the
container to generally have 0 values, 1 when the feature is used, and rarely greater
than 1.

* Source/WebCore/PAL/pal/graphics/WebGPU/Impl/WebGPUDeviceImpl.cpp:
(PAL::WebGPU::DeviceImpl::createBindGroup):

Canonical link: https://commits.webkit.org/264643@main