[WGSL] Missing type promotion for AbstractScalarType https://bugs.webkit.org/show_bug.cgi?id=257380 rdar://109887993
Reviewed by Mike Wyrzykowski.

We currently have a duplicate implementation of Overload::materialize for type
variables. This duplication will be removed soon, but in the mean time we are
missing the promotion step to satisfy constraints in the type variable in one
of them.

* Source/WebGPU/WGSL/Overload.cpp:
(WGSL::OverloadResolver::materialize const):

Canonical link: https://commits.webkit.org/264647@main