[WGSL] Incorrect swizzle validation https://bugs.webkit.org/show_bug.cgi?id=257378 rdar://109886275
Reviewed by Mike Wyrzykowski.

Swizzle characters need to be checked for bounds, e.g. we can't access vec2.z,
but the original patch was incorrectly checking the offset of the swizzle against
the size of the resulting vector, not the vector we are loading from.

* Source/WebGPU/WGSL/TypeCheck.cpp:
(WGSL::TypeChecker::vectorFieldAccess):
* Source/WebGPU/WGSL/tests/invalid/vector.wgsl:
* Source/WebGPU/WGSL/tests/valid/swizzle.wgsl: Added.

Canonical link: https://commits.webkit.org/264650@main