Use signpost instead of trace-point in critical rendering path https://bugs.webkit.org/show_bug.cgi?id=257465 rdar://109977538
Reviewed by NOBODY (OOPS!).

trace-point is *incredibly* costly operation. When it is used in wasm entry point, it caused massive regression[1],
and in JSC, trace-point is used only when explicit option is specified `JSC_useTracePoints`.
And the same issue is happening in WebCore. We have several tracePoint in super critical hot path
like `RenderingUpdateScheduler::scheduleRenderingUpdate` in WebCore, and dominated massive amount of time due to this
system call.

This patch switches them to signpost instead. If we use WTFEmitSignpost etc., it is collected only when explicit option
is attached to the tracing command. This avoids doing super costly tracing in the critical hot path, while we can collect
this information when we would like to do by explicitly attaching an option.

[1]: https://trac.webkit.org/changeset/233378/webkit

* Source/WebCore/page/Page.cpp:
(WebCore::Page::updateRendering):
* Source/WebCore/page/RenderingUpdateScheduler.cpp:
(WebCore::RenderingUpdateScheduler::scheduleRenderingUpdate):
(WebCore::RenderingUpdateScheduler::displayRefreshFired):