Canvas will intermittently come up blurry and remain so. https://bugs.webkit.org/show_bug.cgi?id=256260
Reviewed by Simon Fraser.

The unscaled bitmap optimization uses backing stores that don't have device scaling applied, when the result would be identical.
We need to make sure it gets re-evaluated if the size of the canvas backing store changes, this patch makes sure we schedule a compositing layer update for that to happen.

I haven't found a way to test this yet, capturing the layer tree as text (as other tests of this feature do) implicitly forces a compositing layer update, hiding the bug.

* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::contentChanged):

Canonical link: https://commits.webkit.org/264699@main