REGRESSION (264435@main): Audio player time remainder/spent are oscillating when CSS width has been specified https://bugs.webkit.org/show_bug.cgi?id=257523 rdar://110043258
Reviewed by Aditya Keerthi.

* Source/WebCore/Modules/modern-media-controls/controls/time-control.js:
(TimeControl.prototype._performIdealLayout):
When partially-reverting 264301@main in 264435@main, I re-introduced a
`this.elapsedTimeLabel.visible` which caused this flashing (because of a layout
dependency cycle). Since it is not relevant to the bug I was trying to fix
in 264435@main, remove that condition and make this logic more like the
long-standing code.

Canonical link: https://commits.webkit.org/264760@main